+++ example/lib/src/chroma_counter/views/ui_states/chroma_counter_ui_state.mapper.dart
@@ -1,6 +1,8 @@
 // coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
+// dart format off
 // ignore_for_file: type=lint
+// ignore_for_file: invalid_use_of_protected_member
 // ignore_for_file: unused_element, unnecessary_cast, override_on_non_overriding_member
 // ignore_for_file: strict_raw_type, inference_failure_on_untyped_parameter
 
@@ -21,17 +23,28 @@ class ChromaCounterUIStateMapper extends ClassMapperBase<ChromaCounterUIState> {
   final String id = 'ChromaCounterUIState';
 
   static Color _$backgroundColor(ChromaCounterUIState v) => v.backgroundColor;
-  static const Field<ChromaCounterUIState, Color> _f$backgroundColor =
-      Field('backgroundColor', _$backgroundColor, opt: true, def: Colors.blue);
+  static const Field<ChromaCounterUIState, Color> _f$backgroundColor = Field(
+    'backgroundColor',
+    _$backgroundColor,
+    opt: true,
+    def: Colors.blue,
+  );
   static BorderRadius _$borderRadius(ChromaCounterUIState v) => v.borderRadius;
   static const Field<ChromaCounterUIState, BorderRadius> _f$borderRadius =
       Field('borderRadius', _$borderRadius, opt: true, def: BorderRadius.zero);
   static int _$count(ChromaCounterUIState v) => v.count;
-  static const Field<ChromaCounterUIState, int> _f$count =
-      Field('count', _$count, opt: true, def: 0);
+  static const Field<ChromaCounterUIState, int> _f$count = Field(
+    'count',
+    _$count,
+    opt: true,
+    def: 0,
+  );
   static int _$milestone(ChromaCounterUIState v) => v.milestone;
-  static const Field<ChromaCounterUIState, int> _f$milestone =
-      Field('milestone', _$milestone, mode: FieldMode.member);
+  static const Field<ChromaCounterUIState, int> _f$milestone = Field(
+    'milestone',
+    _$milestone,
+    mode: FieldMode.member,
+  );
 
   @override
   final MappableFields<ChromaCounterUIState> fields = const {
@@ -43,9 +56,10 @@ class ChromaCounterUIStateMapper extends ClassMapperBase<ChromaCounterUIState> {
 
   static ChromaCounterUIState _instantiate(DecodingData data) {
     return ChromaCounterUIState(
-        backgroundColor: data.dec(_f$backgroundColor),
-        borderRadius: data.dec(_f$borderRadius),
-        count: data.dec(_f$count));
+      backgroundColor: data.dec(_f$backgroundColor),
+      borderRadius: data.dec(_f$borderRadius),
+      count: data.dec(_f$count),
+    );
   }
 
   @override
@@ -71,43 +85,57 @@ mixin ChromaCounterUIStateMappable {
         .encodeMap<ChromaCounterUIState>(this as ChromaCounterUIState);
   }
 
-  ChromaCounterUIStateCopyWith<ChromaCounterUIState, ChromaCounterUIState,
-      ChromaCounterUIState> get copyWith => _ChromaCounterUIStateCopyWithImpl<
-          ChromaCounterUIState, ChromaCounterUIState>(
-      this as ChromaCounterUIState, $identity, $identity);
+  ChromaCounterUIStateCopyWith<
+    ChromaCounterUIState,
+    ChromaCounterUIState,
+    ChromaCounterUIState
+  >
+  get copyWith =>
+      _ChromaCounterUIStateCopyWithImpl<
+        ChromaCounterUIState,
+        ChromaCounterUIState
+      >(this as ChromaCounterUIState, $identity, $identity);
   @override
   String toString() {
-    return ChromaCounterUIStateMapper.ensureInitialized()
-        .stringifyValue(this as ChromaCounterUIState);
+    return ChromaCounterUIStateMapper.ensureInitialized().stringifyValue(
+      this as ChromaCounterUIState,
+    );
   }
 
   @override
   bool operator ==(Object other) {
-    return ChromaCounterUIStateMapper.ensureInitialized()
-        .equalsValue(this as ChromaCounterUIState, other);
+    return ChromaCounterUIStateMapper.ensureInitialized().equalsValue(
+      this as ChromaCounterUIState,
+      other,
+    );
   }
 
   @override
   int get hashCode {
-    return ChromaCounterUIStateMapper.ensureInitialized()
-        .hashValue(this as ChromaCounterUIState);
+    return ChromaCounterUIStateMapper.ensureInitialized().hashValue(
+      this as ChromaCounterUIState,
+    );
   }
 }
 
 extension ChromaCounterUIStateValueCopy<$R, $Out>
     on ObjectCopyWith<$R, ChromaCounterUIState, $Out> {
   ChromaCounterUIStateCopyWith<$R, ChromaCounterUIState, $Out>
-      get $asChromaCounterUIState => $base.as(
-          (v, t, t2) => _ChromaCounterUIStateCopyWithImpl<$R, $Out>(v, t, t2));
+  get $asChromaCounterUIState => $base.as(
+    (v, t, t2) => _ChromaCounterUIStateCopyWithImpl<$R, $Out>(v, t, t2),
+  );
 }
 
 abstract class ChromaCounterUIStateCopyWith<
-    $R,
-    $In extends ChromaCounterUIState,
-    $Out> implements ClassCopyWith<$R, $In, $Out> {
+  $R,
+  $In extends ChromaCounterUIState,
+  $Out
+>
+    implements ClassCopyWith<$R, $In, $Out> {
   $R call({Color? backgroundColor, BorderRadius? borderRadius, int? count});
   ChromaCounterUIStateCopyWith<$R2, $In, $Out2> $chain<$R2, $Out2>(
-      Then<$Out2, $R2> t);
+    Then<$Out2, $R2> t,
+  );
 }
 
 class _ChromaCounterUIStateCopyWithImpl<$R, $Out>
@@ -120,19 +148,23 @@ class _ChromaCounterUIStateCopyWithImpl<$R, $Out>
       ChromaCounterUIStateMapper.ensureInitialized();
   @override
   $R call({Color? backgroundColor, BorderRadius? borderRadius, int? count}) =>
-      $apply(FieldCopyWithData({
-        if (backgroundColor != null) #backgroundColor: backgroundColor,
-        if (borderRadius != null) #borderRadius: borderRadius,
-        if (count != null) #count: count
-      }));
+      $apply(
+        FieldCopyWithData({
+          if (backgroundColor != null) #backgroundColor: backgroundColor,
+          if (borderRadius != null) #borderRadius: borderRadius,
+          if (count != null) #count: count,
+        }),
+      );
   @override
   ChromaCounterUIState $make(CopyWithData data) => ChromaCounterUIState(
-      backgroundColor: data.get(#backgroundColor, or: $value.backgroundColor),
-      borderRadius: data.get(#borderRadius, or: $value.borderRadius),
-      count: data.get(#count, or: $value.count));
+    backgroundColor: data.get(#backgroundColor, or: $value.backgroundColor),
+    borderRadius: data.get(#borderRadius, or: $value.borderRadius),
+    count: data.get(#count, or: $value.count),
+  );
 
   @override
   ChromaCounterUIStateCopyWith<$R2, ChromaCounterUIState, $Out2>
-      $chain<$R2, $Out2>(Then<$Out2, $R2> t) =>
-          _ChromaCounterUIStateCopyWithImpl<$R2, $Out2>($value, $cast, t);
+  $chain<$R2, $Out2>(Then<$Out2, $R2> t) =>
+      _ChromaCounterUIStateCopyWithImpl<$R2, $Out2>($value, $cast, t);
 }
+